Improve how forms and validation work in Blazor web apps module completed

+++ BlazorApp1/Models/Pizza.cs
@@ -22,7 +22,7 @@ public class Pizza
 
     public decimal GetBasePrice()
     {
-        return ((decimal)Size / (decimal)DefaultSize) * Special.BasePrice;
+        return Size / (decimal)DefaultSize * Special.BasePrice;
     }
 
     public decimal GetTotalPrice()
